% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{jasco_boxplot}
\alias{jasco_boxplot}
\title{jasco_boxplot}
\usage{
jasco_boxplot(df, x, ...)
}
\arguments{
\item{df}{data to plot}

\item{x}{x-axis}

\item{a}{facet_v}

\item{b}{facet_h}
}
\description{
fast n' easy default plot!
}
\examples{
jasco_boxplot()
}
\keyword{jasco}
