% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{jasco_extract_lm}
\alias{jasco_extract_lm}
\title{jasco_extract_lm}
\usage{
jasco_extract_lm(data, response, predictor, min, max, ...)
}
\arguments{
\item{data}{a tibble}

\item{response}{unquoted name of the response variable found in the tibble}

\item{predictor}{unquoted name of the response variable found in the tibble}
}
\description{
This is a wrapper for lm and a helper function for jasco_extract_lm.
}
\details{
For a given tibble, this function allows to ...
}
\examples{
jasco_extract_lm()
}
\keyword{jasco}
