% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{jasco_extract_params}
\alias{jasco_extract_params}
\title{jasco_extract_params}
\usage{
jasco_extract_params(.data, .bgcorr, .protein_uM, .unit, .factor, ...)
}
\arguments{
\item{.data}{a tibble}

\item{.bgcorr}{logical}

\item{.predictor}{unquoted name of the response variable found in the tibble}
}
\description{
sentence here
}
\details{
For a given tibble, this function allows to ...
}
\examples{
jasco_extract_params()
}
\keyword{jasco}
