% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{jasco_get_turnover}
\alias{jasco_get_turnover}
\title{jasco_get_turnover}
\usage{
jasco_get_turnover(.rate, .protein_uM, .unit, .factor)
}
\arguments{
\item{.protein_uM}{final protein concentration in assay}

\item{.unit}{current time scale; one of h = hour, min = minutes, sec = seconds, msec = milliseconds}

\item{.factor}{coupling factor; eg. for ATPase assay NADH decline is monitored:
each oxidized NADH molecule reports the hydrolysis of one molecule of ATP; hence, coupling factor is `-1`}

\item{.data}{a tibble, minimal column `rate_avg`}
}
\description{
sentence here
}
\details{
For a given tibble, this function allows to ...
}
\examples{
jasco_get_turnover()
}
\keyword{jasco}
