% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{jasco_read}
\alias{jasco_read}
\title{jasco_read}
\usage{
jasco_read(file)
}
\arguments{
\item{file}{full path to .txt file}
}
\description{
This is a function for easy import of data from the Spectrophotometer (JASCO Corp., V-560, Rev. 1.00)
}
\details{
Output is a tidy tibble containing a Set_ID (date-time of record), an Exp_ID (basename of provided file),
the instrument name and the actual data, which is a Absorbance reading `Absorbance` as a function of time `Time_s`.
}
\examples{
file <- system.file("extdata", "1.txt" , package = "jasco2", mustWork = TRUE)
jasco_read(file)
}
\keyword{jasco}
