% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\docType{data}
\name{my_theme_addclassic_boxplot}
\alias{my_theme_addclassic_boxplot}
\title{my_theme_addclassic_boxplot}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 12.
}
\usage{
my_theme_addclassic_boxplot
}
\description{
my_theme_addclassic_boxplot
}
\keyword{datasets}
