% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_absorbance.R
\name{convert_absorbance}
\alias{convert_absorbance}
\title{convert_absorbance}
\usage{
convert_absorbance(data, coef)
}
\arguments{
\item{data}{a tibble with at least one column 'Absorbance'.}

\item{coef}{either the name of any chemical found in `chemicals_specs` or a
user-provided appropriate absorption coefficient (a named vector, eg. `setNames(7282, "chemical1")`)}
}
\description{
given a tibble containing a column 'Absorbance', the concentration of a chemical
  of which the Absorbance is measured is calculated and attached to the tibble.
  Userprovided coefficients return the calculated concentration in corresponding
  units - If the coefficient is one of the chemicals provided in 'chemical_specs.R'
  the calculated concentration is stated in mol/l .
}
\examples{

# this is a typical output from jasco_tibble
jasco_df

jasco_df \%>\%
  convert_absorbance(., "NADH")
  
jasco_df \%>\%
  convert_absorbance(., coef = setNames(7282, "chemical XYZ"))
}
\keyword{jasco}
