% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{jasco_extract_lm}
\alias{jasco_extract_lm}
\title{jasco_extract_lm}
\usage{
jasco_extract_lm(data, response, predictor, min, max, ...)
}
\arguments{
\item{data}{a tibble with spectrophotometer data, eg. output from jasco_tibble}

\item{response}{unquoted name of the response variable found in the tibble}

\item{predictor}{unquoted name of the predictor variable found in the tibble}
}
\value{
tibble, that includes parameters and statistics of the fitted linear 
  model using broom::glance, broom::tidy and broom::augment, nested per experiment.
  Metadata stated per experiment is maintained.
}
\description{
This is a wrapper for lm and a helper function for jasco_extract_lm.
}
\details{
For a given tibble, this function allows to ...
}
\examples{

# this is a typical output from jasco_tibble
jasco_df

jasco_df \%>\%
  jasco_extract_lm(., response = Absorbance, predictor = Time_s, min = 75, max = 175)

# Conversion from Absorbance to NADH concentration (mol/l) is optional
jasco_df \%>\%
  convert_absorbance(., "NADH") \%>\%
  jasco_extract_lm(., response = NADH, predictor = Time_s, min = 75, max = 175)
}
\keyword{jasco}
