% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{jasco_extract_params}
\alias{jasco_extract_params}
\title{jasco_extract_params}
\usage{
jasco_extract_params(.data, .bgcorr, .protein_uM, .unit, .factor, ...)
}
\arguments{
\item{.data}{a tibble, minimal column `rate_avg`}

\item{.bgcorr}{optional background correction (subtraction); NADH oxidation in absence of protein}

\item{.protein_uM}{final protein concentration in assay}

\item{.unit}{current time scale; one of h = hour, min = minutes, sec = seconds, msec = milliseconds}

\item{.factor}{coupling factor; eg. for ATPase assay NADH decline is monitored:
each oxidized NADH molecule reports the hydrolysis of one molecule of ATP; hence, coupling factor is `-1`}

\item{.group_vars}{grouping variable for summary report (eg. 'Group' - for grouping replicates)}
}
\value{
list of two tibbles; 'data' and 'summary'. 'data' contains all columns,
  including model statistics for evaluation. 'summary' is a grouped summary of
  turnover(min^-1) ('turnover_avg') and sd ('turnover_sd')
}
\description{
extracts turnover of ATP molecules per protein (min^-1). For input, see examples or README.
}
\examples{

# typical output from jasco_tibble()
jasco_df

jasco_df \%>\%
  convert_absorbance(., "NADH") \%>\%
  jasco_extract_lm(., response = NADH, predictor = Time_s, min = 75, max = 175) \%>\%
  jasco_extract_params(., .bgcorr = FALSE, .protein_uM = 6, .unit = "sec", .factor = -1, Group)
}
\keyword{jasco}
